Refactor tests for watched addressess methods
